# docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: Bosch Car Service Companion API
  version: 0.1.0
  description: >
    API REST pentru aplicația client (PWA) și portalul service (single-tenant).
    Toate mesajele orientate spre utilizator sunt în limba română.
    Erorile folosesc un format standardizat (application/problem+json).
servers:
  - url: https://localhost/api
    description: Local (Nginx + HTTPS)
...
paths:
...
  /documents:
    post:
      tags: [Documents]
      summary: Încarcă document (multipart; validare MIME+extensie+dimensiune, scanare async)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        '201':
          description: Creat (scanStatus inițial PENDING)
          content:
            application/json:
              schema: { $ref: '#/components/schemas/DocumentMeta' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /documents/{id}/download-url:
    parameters: [ { name: id, in: path, required: true, schema: { type: string, format: uuid } } ]
    get:
      tags: [Documents]
      summary: URL temporar semnat de descărcare (doar documente CLEAN, autorizat)
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  url: { type: string }
                  expiresAt: { type: string, format: date-time }
        '403': { $ref: '#/components/responses/Error' }
        '422': { $ref: '#/components/responses/ValidationError' }
  /documents/{id}/raw:
    parameters:
      - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      - { name: expires, in: query, required: true, schema: { type: integer } }
      - { name: sig, in: query, required: true, schema: { type: string } }
    get:
      tags: [Documents]
      summary: Servire conținut (semnătură + expirare + autorizare la nivel de obiect)
      responses:
        '200': { description: Conținut binar }
        '403': { $ref: '#/components/responses/Error' }
        '404': { $ref: '#/components/responses/Error' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: BCSC_SESSION
...
  responses:
    Error:
      description: Eroare
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
    ValidationError:
      description: Date invalide
      content:
        application/problem+json:
          schema: { $ref: '#/components/schemas/Problem' }
          example:
            type: validation_error
            title: Date invalide
            status: 422
            errors:
              expiresAt: ["Data expirării trebuie să fie ulterioară datei de început."]
            traceId: "01J..."
  schemas:
    Problem:
      type: object
      required: [type, title, status, traceId]
      properties:
        type: { type: string, example: validation_error }
        title: { type: string, example: Date invalide }
        status: { type: integer, example: 422 }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
        traceId: { type: string }
...
    DocumentMeta:
      type: object
      properties:
        id: { type: string, format: uuid }
        mimeType: { type: string }
        sizeBytes: { type: integer }
        scanStatus: { type: string, enum: [PENDING, CLEAN, INFECTED] }
security:
  - sessionCookie: []